% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkps_method.R
\docType{methods}
\name{brkps}
\alias{brkps}
\alias{brkps,MSigSeg-method}
\title{Generic Function-brkps.}
\usage{
\S4method{brkps}{MSigSeg}(object)
}
\arguments{
\item{object}{A 'MSigSeg' object.}
}
\value{
The brkps slot of MSigSeg object.
}
\description{
This function returns the brkps slot of MSigSeg object.
}
\details{
This function is a S4 method for MSigSeg object. It retrieves brkps slot, which contains the locations of break points.
}
\examples{
x=new("MSigSeg") # Creating a new MSigSeg object.
brkps(x)

}
