% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change}
\alias{change}
\title{Breakpoints matrix generation.}
\usage{
change(M, p = 0.01)
}
\arguments{
\item{M}{A matrix users aimed to add breakpoints.}

\item{p}{Probability of occurrence of breakpoints.}
}
\value{
A list containing the matrix with specified change point and the location of breakpoints.
}
\description{
Generate matrix based on specified breakpoints.
}
\details{
Generate matrix with common breakpoints, based on specific probability of occurrence.
}
