% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.input_method.R
\docType{methods}
\name{data.input}
\alias{data.input}
\alias{data.input,MSigSeg-method}
\title{Generic Function-data.input.}
\usage{
\S4method{data.input}{MSigSeg}(object)
}
\arguments{
\item{object}{A MSigSeg object.}
}
\value{
The data.input slot of MSigSeg object.
}
\description{
This function returns the data.input slot of MSigSeg object.
}
\details{
This function is a S4 method for MSigSeg object. It retrieves data.input slot, which contains the data users input.
}
\examples{
x=new("MSigSeg") # Creating a new MSigSeg object.
data.input(x)

}
