% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.output_method.R
\docType{methods}
\name{data.output}
\alias{data.output}
\alias{data.output,MSigSeg-method}
\title{Generic Function-data.output.}
\usage{
\S4method{data.output}{MSigSeg}(object)
}
\arguments{
\item{object}{A MSigSeg object.}
}
\value{
The data.output slot of MSigSeg object.
}
\description{
This function returns the data.output slot of MSigSeg object.
}
\details{
This function is a S4 method for MSigSeg object. It retrieves data.output slot, which contains the input data which has been smoothed..
}
\examples{
x=new("MSigSeg") # Creating a new MSigSeg object.
data.output(x)

}
