% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_test}
\alias{data_test}
\title{A simulated data set used for testing.}
\format{
A matrix with 1000 rows and 20 columns.
}
\usage{
data_test
}
\description{
A simulated data set used for testing.
}
\examples{
data("data_test",package = "MSigSeg")
}
\keyword{datasets}
