% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_change.R
\name{fun_change}
\alias{fun_change}
\title{fun_change.}
\usage{
fun_change(M, p = 0.01)
}
\arguments{
\item{M}{A matrix users aimed to add change points.}

\item{p}{Probability of occurrence of change points.}
}
\value{
A list containing the matrix with specified change point and the location of change points.
}
\description{
Generate matrix based on specified change points.
}
\details{
Generate matrix with common change points, based on specific probability of occurrence.
}
