% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_lambda_estimator.R
\name{fun_lambda_estimator}
\alias{fun_lambda_estimator}
\title{Detecting common change points with designated number.}
\usage{
fun_lambda_estimator(Y, K)
}
\arguments{
\item{Y}{An data.frame/matrix containing the data within which you wish to find common change points. Each column is considered a separate signal.}

\item{K}{Number of change points users want to detect.}
}
\value{
An object of S4 class "MSigSeg".
}
\description{
Calculates the optimal positioning of common change points with designated number.
}
\details{
This function is based on the fun_segmentation_PELT() function. Number of change points are defined by users and lambda is calculated by algorithm automatically.
}
\examples{
data(PELT_test)
fun_lambda_estimator(PELT_test,5)

}
