% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_method.R
\docType{methods}
\name{lambda}
\alias{lambda}
\alias{lambda,MSigSeg-method}
\title{Generic Function-lambda.}
\usage{
\S4method{lambda}{MSigSeg}(object)
}
\arguments{
\item{object}{A MSigSeg object.}
}
\value{
The lambda slot of MSigSeg object.
}
\description{
This function returns the lambda slot of MSigSeg object.
}
\details{
This function is a S4 method for MSigSeg object. It retrieves lambda slot, which contains penalty coefficient to prevent over fitting.
}
\examples{
x=new("MSigSeg") # Creating a new MSigSeg object.
lambda(x)

}
