% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_estimator.R
\name{lambda_estimator}
\alias{lambda_estimator}
\title{Detecting common breakpoints with designated number.}
\usage{
lambda_estimator(Y, K)
}
\arguments{
\item{Y}{An data.frame/matrix containing the data to be segmented. Each column stores a signal.}

\item{K}{Number of change points users want to detect.}
}
\value{
An object of S4 class "MSigSeg".
}
\description{
Automatic estimation of penalty parameter lambda for user defined breakpoints number.
}
\details{
This function is based on the segmentation() function. Number of breakpoints are defined by users and lambda is calculated by algorithm automatically.
}
\examples{
data(data_test)
lambda_estimator(data_test,5)

}
