% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSigSeg_Class.R
\docType{class}
\name{MSigSeg-class}
\alias{MSigSeg-class}
\title{An S4 class to encapsulation the result of changepoint analysis.}
\description{
An S4 class to encapsulation the result of changepoint analysis.
}
\section{Slots}{

\describe{
\item{\code{data.input}}{An data.frame/matrix containing the data within which you wish to find common change points. Each column is considered a separate signal.}

\item{\code{data.output}}{A matrix containing the input data which has been smoothed.}

\item{\code{lambda}}{A numeric containing penalty coefficient to prevent over fitting.}

\item{\code{brkps}}{A vector containing the locations of common change points.}

\item{\code{fmin}}{A numeric containing the optimal numerical value calculated.}

\item{\code{date}}{Character string containing date information.}
}}

