% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_plot.R
\name{multi_plot}
\alias{multi_plot}
\title{Plot function of MSigSeg package.}
\usage{
multi_plot(m, ncol, nrow)
}
\arguments{
\item{m}{An object of S4 class "MSigSeg".}

\item{ncol}{Column numbers of signals arrangement in the graph.}

\item{nrow}{Row numbers of signals arrangement in the graph}
}
\value{
A list, first item in the list is a graphic objects with all signals drawn and second is a list with individual signals.
}
\description{
Graph signals and breakpoints based on ggplot2 and ggarange packages.
}
\examples{
data(data_test)
m <- segmentation(data_test,100)
p <- multi_plot(m,4,5)

}
