% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCHSData}
\alias{NCHSData}
\title{influenza  data set from CDC used as an example.}
\format{
A matrix with 52 rows and 10 columns.
}
\usage{
NCHSData
}
\description{
influenza  data set from CDC used as an example.
}
\examples{
data("NCHSData",package = "MSigSeg")
}
\keyword{datasets}
