% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisegen.R
\name{noisegen}
\alias{noisegen}
\title{Noisegen.}
\usage{
noisegen(X, SNR)
}
\arguments{
\item{X}{A matrix users aimed to add signal-to-noise ratio.}

\item{SNR}{Signal-to-noise ratio.}
}
\value{
A matrix with specified signal-to-noise ratio.
}
\description{
Generate matrix based on signal-to-noise ratio.
}
