% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PELT_test}
\alias{PELT_test}
\title{A data set used for testing.}
\format{
A matrix with 1000 rows and 20 columns.
}
\usage{
PELT_test
}
\description{
A data set used for testing.
}
\keyword{datasets}
