% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_method.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,MSigSeg-method}
\title{Generic Function-print.}
\usage{
\S4method{print}{MSigSeg}(object)
}
\arguments{
\item{object}{A MSigSeg object.}
}
\value{
The the basic information of MSigSeg object.
}
\description{
This function print the basic information of MSigSeg object.
}
\details{
This function is a S4 method for MSigSeg object. It prints class, slots, created  date and summary of MSigSeg object.
}
\examples{
x=new("MSigSeg") # Creating a new MSigSeg object.
print(x)

}
