% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg.len_method.R
\docType{methods}
\name{seg.len}
\alias{seg.len}
\alias{seg.len,MSigSeg-method}
\title{Generic Function-seg.len.}
\usage{
\S4method{seg.len}{MSigSeg}(object)
}
\arguments{
\item{object}{A MSigSeg object.}
}
\value{
A vector contains length of segmentation.
}
\description{
This function returns the length of segmentation.
}
\details{
This function is a S4 method for MSigSeg object. It calculates the distance between each change points.
}
\examples{
x=new("MSigSeg") # Creating a new MSigSeg object.
seg.len(x)

}
