% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock}
\alias{stock}
\title{A stock data set used as an example.}
\format{
A data.frame with 757 rows and 488 columns.
}
\usage{
stock
}
\description{
A stock data set used as an example.
}
\examples{
data("stock",package = "MSigSeg")
}
\keyword{datasets}
