% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_method.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,MSigSeg-method}
\title{Generic Function-summary.}
\usage{
\S4method{summary}{MSigSeg}(object)
}
\arguments{
\item{object}{A MSigSeg object.}
}
\value{
A summary of MSigSeg object.
}
\description{
This function summarize the information of MSigSeg object.
}
\details{
This function is a S4 method for MSigSeg object. It summarizes the number of signals, length of signals, number of change points and fmin.
}
\examples{
x=new("MSigSeg") # Creating a new MSigSeg object.
summary(x)

}
