% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T16P}
\alias{T16P}
\title{A chromosome sequencing data set used as an example.}
\format{
A data.frame with 2928 rows and 16 columns.
}
\usage{
T16P
}
\description{
A chromosome sequencing data set used as an example.
}
\examples{
data("T16P",package = "MSigSeg")
}
\references{
Navin N, Kendall J, Troge J, et al. Tumour evolution inferred by single-cell sequencing. Nature. 2011;472(7341):90-94. doi:10.1038/nature09807
}
\keyword{datasets}
