% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSigSeg_Class.R
\docType{class}
\name{MSigSeg-class}
\alias{MSigSeg-class}
\title{An S4 class to encapsulation the result of breakpoints analysis.}
\description{
An S4 class to encapsulation the result of breakpoints analysis.
}
\section{Slots}{

\describe{
\item{\code{data.input}}{An data.frame/matrix containing the data to be segmented. Each column stores a signal.}

\item{\code{data.output}}{A matrix containing the input data which has been smoothed.}

\item{\code{lambda}}{A penalty term, small value leads to large number of breakpoints, and vice versa.}

\item{\code{brkps}}{A vector containing the locations of common breakpoints.}

\item{\code{fmin}}{A numeric containing the optimal numerical value calculated.}

\item{\code{date}}{Character string containing date information.}
}}

